Split permit model class file
